backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
supabase==2.3.3
python-dotenv==1.0.1
